tests: add helper for executing a checking the exit code of a command
The execCmd test helper is refactored to:
- accept any cobra.Command
- fail the test also if the command succeeds but is expected to fail
- be named execCheck
